Refactored and organized
