+++ src/styles.css
@@ -74,7 +74,7 @@
 .rt-content h1, h2, h3 {
   margin-top: 0;
   margin-bottom: 8px;
-  font-weight: 600;
+  font-weight: 400;
 }
 
 .rt-content h1 {
