+++ src/utils/index.ts
@@ -5,3 +5,9 @@ export { setCursorAtStartOfElement } from "./setCursorAtStartOfElement";
 export { removeTagFromSelection } from "./removeTagFromSelection";
 export { sanitizeContent } from "./sanitizeContent";
 export { setCursorAtStart } from "./setCursorAtStart";
+export { handleChange } from "./handleChange";
+export { toggleStyle } from "./toggleStyle";
+export { toggleListStyle } from "./toggleListStyle";
+export { toggleHeaderStyle } from "./toggleHeaderStyle";
+export { removeHeadersFromSelection } from "./removeHeadersFromSelection";
+export { handleKeyDown } from "./handleKeyDown";
